\name{codareg-package}
\alias{codareg-package}
\alias{codareg}
\docType{package}
\title{
\packageTitle{codareg}
}
\description{
\packageDescription{codareg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{codareg}
\packageIndices{codareg}
}
\author{
\packageAuthor{codareg}

Maintainer: \packageMaintainer{codareg}
}
\references{
Morais, J., & Thomas-Agnan, C. (2020). Covariates impacts in compositional models and simplicial derivatives. Austrian Journal of Statistics.
}

\keyword{ package }

